+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## TBD
+### Changed
+* Default Housekeeping configuration is now provided via `@Bean` `housekeepingEnvironment` only if the bean is not already provided.
+
 ## [1.0.2] 2018-02-15
 ### Added
 * Added new housekeeping configuration parameters to set the schema name and the location of the DB initialization script.